GPU: Merge opencl-common and opencl2 source files
