--- GPU/GPUTracking/Base/opencl-common/CMakeLists.txt
@@ -1,56 +1,0 @@
-# Copyright 2019-2020 CERN and copyright holders of ALICE O2.
-# See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
-# All rights not expressly granted are reserved.
-#
-# This software is distributed under the terms of the GNU General Public
-# License v3 (GPL Version 3), copied verbatim in the file "COPYING".
-#
-# In applying this license CERN does not waive the privileges and immunities
-# granted to it by virtue of its status as an Intergovernmental Organization
-# or submit itself to any jurisdiction.
-
-set(MODULE GPUTrackingOpenCLCommon)
-
-set(SRCS GPUReconstructionOCL.cxx)
-set(HDRS GPUReconstructionOCL.h GPUReconstructionOCLInternals.h)
-
-if(ALIGPU_BUILD_TYPE STREQUAL "O2")
-  o2_add_library(${MODULE}
-                 SOURCES ${SRCS}
-                 PUBLIC_INCLUDE_DIRECTORIES ${CMAKE_CURRENT_LIST_DIR}
-                 PUBLIC_LINK_LIBRARIES OpenCL::OpenCL O2::GPUTracking
-                 TARGETVARNAME targetName)
-
-  target_compile_definitions(${targetName} PRIVATE $<TARGET_PROPERTY:O2::GPUTracking,COMPILE_DEFINITIONS>)
-  # the compile_defitions are not propagated automatically on purpose (they are
-  # declared PRIVATE) so we are not leaking them outside of the GPU**
-  # directories
-
-  install(FILES ${HDRS} DESTINATION include/GPU)
-endif()
-
-if(ALIGPU_BUILD_TYPE STREQUAL "ALIROOT")
-  # Generate the dictionary
-  get_directory_property(incdirs INCLUDE_DIRECTORIES)
-  generate_dictionary("Ali${MODULE}" "" "GPUReconstructionOCL.h" "${incdirs} .")
-
-  # Generate the ROOT map
-  generate_rootmap("Ali${MODULE}" "" "")
-
-  # Add a library to the project using the specified source files
-  add_library_tested(Ali${MODULE} SHARED ${SRCS} G__Ali${MODULE}.cxx)
-  # AMD OpenCL run-time and driver
-  target_link_libraries(Ali${MODULE} PUBLIC OpenCL AliGPUTracking)
-
-  # Installation
-  install(TARGETS Ali${MODULE} ARCHIVE DESTINATION lib LIBRARY DESTINATION lib)
-
-  install(FILES ${HDRS} DESTINATION include)
-endif()
-
-if(ALIGPU_BUILD_TYPE STREQUAL "Standalone")
-  add_library(${MODULE} SHARED ${SRCS})
-  target_link_libraries(${MODULE} GPUTracking OpenCL)
-  target_include_directories(${MODULE} PUBLIC ${CMAKE_CURRENT_LIST_DIR})
-  install(TARGETS ${MODULE})
-endif()
