--- GPU/GPUTracking/Base/opencl2/GPUReconstructionOCL2.cxx
@@ -1,144 +1,0 @@
-// Copyright 2019-2020 CERN and copyright holders of ALICE O2.
-// See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
-// All rights not expressly granted are reserved.
-//
-// This software is distributed under the terms of the GNU General Public
-// License v3 (GPL Version 3), copied verbatim in the file "COPYING".
-//
-// In applying this license CERN does not waive the privileges and immunities
-// granted to it by virtue of its status as an Intergovernmental Organization
-// or submit itself to any jurisdiction.
-
-/// \file GPUReconstructionOCL.cxx
-/// \author David Rohr
-
-#define GPUCA_GPUTYPE_OPENCL
-#define __OPENCL_HOST__
-
-#include "GPUReconstructionOCL2.h"
-#include "GPUReconstructionOCLInternals.h"
-#include "GPUReconstructionIncludes.h"
-
-using namespace GPUCA_NAMESPACE::gpu;
-
-#include <cstring>
-#include <unistd.h>
-#include <typeinfo>
-#include <cstdlib>
-
-#include "utils/qGetLdBinarySymbols.h"
-QGET_LD_BINARY_SYMBOLS(GPUReconstructionOCLCode_src);
-#ifdef OPENCL_ENABLED_SPIRV
-QGET_LD_BINARY_SYMBOLS(GPUReconstructionOCLCode_spirv);
-#endif
-
-GPUReconstruction* GPUReconstruction_Create_OCL(const GPUSettingsDeviceBackend& cfg) { return new GPUReconstructionOCL2(cfg); }
-
-GPUReconstructionOCLBackend::GPUReconstructionOCLBackend(const GPUSettingsDeviceBackend& cfg) : GPUReconstructionOCL(cfg)
-{
-}
-
-template <class T, int32_t I, typename... Args>
-int32_t GPUReconstructionOCLBackend::runKernelBackend(const krnlSetupArgs<T, I, Args...>& args)
-{
-  cl_kernel k = args.s.y.num > 1 ? getKernelObject<cl_kernel, T, I, true>() : getKernelObject<cl_kernel, T, I, false>();
-  return std::apply([this, &args, &k](auto&... vals) { return runKernelBackendInternal(args.s, k, vals...); }, args.v);
-}
-
-template <class S, class T, int32_t I, bool MULTI>
-S& GPUReconstructionOCLBackend::getKernelObject()
-{
-  static uint32_t krnl = FindKernel<T, I>(MULTI ? 2 : 1);
-  return mInternals->kernels[krnl].first;
-}
-
-int32_t GPUReconstructionOCLBackend::GetOCLPrograms()
-{
-  char platform_version[256] = {};
-  GPUFailedMsg(clGetPlatformInfo(mInternals->platform, CL_PLATFORM_VERSION, sizeof(platform_version), platform_version, nullptr));
-  float ver = 0;
-  sscanf(platform_version, "OpenCL %f", &ver);
-
-  cl_int ocl_error;
-
-  const char* ocl_flags = GPUCA_M_STR(OCL_FLAGS);
-
-#ifdef OPENCL_ENABLED_SPIRV // clang-format off
-  if (ver >= 2.2f && !GetProcessingSettings().oclCompileFromSources) {
-    GPUInfo("Reading OpenCL program from SPIR-V IL (Platform version %4.2f)", ver);
-    mInternals->program = clCreateProgramWithIL(mInternals->context, _binary_GPUReconstructionOCLCode_spirv_start, _binary_GPUReconstructionOCLCode_spirv_len, &ocl_error);
-    ocl_flags = "";
-  } else
-#endif // clang-format on
-  {
-    GPUInfo("Compiling OpenCL program from sources (Platform version %4.2f)", ver);
-    size_t program_sizes[1] = {_binary_GPUReconstructionOCLCode_src_len};
-    char* programs_sources[1] = {_binary_GPUReconstructionOCLCode_src_start};
-    mInternals->program = clCreateProgramWithSource(mInternals->context, (cl_uint)1, (const char**)&programs_sources, program_sizes, &ocl_error);
-  }
-
-  if (GPUFailedMsgI(ocl_error)) {
-    GPUError("Error creating OpenCL program from binary");
-    return 1;
-  }
-
-  if (GPUFailedMsgI(clBuildProgram(mInternals->program, 1, &mInternals->device, ocl_flags, nullptr, nullptr))) {
-    cl_build_status status;
-    if (GPUFailedMsgI(clGetProgramBuildInfo(mInternals->program, mInternals->device, CL_PROGRAM_BUILD_STATUS, sizeof(status), &status, nullptr)) == 0 && status == CL_BUILD_ERROR) {
-      size_t log_size;
-      clGetProgramBuildInfo(mInternals->program, mInternals->device, CL_PROGRAM_BUILD_LOG, 0, nullptr, &log_size);
-      std::unique_ptr<char[]> build_log(new char[log_size + 1]);
-      clGetProgramBuildInfo(mInternals->program, mInternals->device, CL_PROGRAM_BUILD_LOG, log_size, build_log.get(), nullptr);
-      build_log[log_size] = 0;
-      GPUError("Build Log:\n\n%s\n", build_log.get());
-    }
-    return 1;
-  }
-
-#define GPUCA_KRNL(...) \
-  GPUCA_KRNL_WRAP(GPUCA_KRNL_LOAD_, __VA_ARGS__)
-#define GPUCA_KRNL_LOAD_single(x_class, ...)              \
-  if (AddKernel<GPUCA_M_KRNL_TEMPLATE(x_class)>(false)) { \
-    return 1;                                             \
-  }
-#define GPUCA_KRNL_LOAD_multi(x_class, ...)              \
-  if (AddKernel<GPUCA_M_KRNL_TEMPLATE(x_class)>(true)) { \
-    return 1;                                            \
-  }
-#include "GPUReconstructionKernelList.h"
-#undef GPUCA_KRNL
-#undef GPUCA_KRNL_LOAD_single
-#undef GPUCA_KRNL_LOAD_multi
-
-  return 0;
-}
-
-bool GPUReconstructionOCLBackend::CheckPlatform(uint32_t i)
-{
-  char platform_version[64] = {}, platform_vendor[64] = {};
-  clGetPlatformInfo(mInternals->platforms[i], CL_PLATFORM_VERSION, sizeof(platform_version), platform_version, nullptr);
-  clGetPlatformInfo(mInternals->platforms[i], CL_PLATFORM_VENDOR, sizeof(platform_vendor), platform_vendor, nullptr);
-  float ver1 = 0;
-  sscanf(platform_version, "OpenCL %f", &ver1);
-  if (ver1 >= 2.2f) {
-    if (mProcessingSettings.debugLevel >= 2) {
-      GPUInfo("OpenCL 2.2 capable platform found");
-    }
-    return true;
-  }
-
-  if (strcmp(platform_vendor, "Advanced Micro Devices, Inc.") == 0 && ver1 >= 2.0f) {
-    float ver2 = 0;
-    const char* pos = strchr(platform_version, '(');
-    if (pos) {
-      sscanf(pos, "(%f)", &ver2);
-    }
-    if ((ver1 >= 2.f && ver2 >= 2000.f) || ver1 >= 2.1f) {
-      if (mProcessingSettings.debugLevel >= 2) {
-        GPUInfo("AMD ROCm OpenCL Platform found");
-      }
-      return true;
-    }
-  }
-  return false;
-}
